% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmip6_center_scale}
\alias{cmip6_center_scale}
\title{A list of the CMIP6 scenario, variable, and year center and scale weights.}
\format{A list of the center and scale values to normalize Hector and ESM output data.
\describe{
\item{center}{A vector the Hector ensemble means for each scenario, variable, and year to use when noramlizing data.}
\item{scale}{A vector the Hector ensemble standard deviation for each scenario, variable, and year to use when noramlizing data.}
}}
\usage{
cmip6_center_scale
}
\description{
A list of the values from the large ensemble of Hector values to use to
center and scale the Hector and CMIP6 output by. This is an important
part of the single ESM calibration process.
}
\seealso{
Other CMIP6: \code{\link{cmip6_Hector_inputs}},
  \code{\link{cmip6_data}}, \code{\link{cmip6_ini}}
}
\concept{CMIP6}
\keyword{datasets}
