% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmip6_data}
\alias{cmip6_data}
\title{A data frame of the global mean CMIP6 values}
\format{Data frame with 7 columns
\describe{
\item{era}{The MIP era}
\item{year}{Year being described (1851--2300)}
\item{value}{Value of the variable being reported}
\item{variable}{Variable being reported ("tas" or "co2").  Temperatures are
anomalies relative to the first decade reported by the model.}
\item{model}{Name of the model}
\item{ensemble}{Identifier for the ensemble member}
\item{experiment}{Experiment being reported. One of "historical", "rcp26", "rcp45",
"rcp60", "rcp85", "esmHistorical", "esmrcp85"}
}}
\usage{
cmip6_data
}
\description{
The data frame containing the global mean value from each ESM to use as comparison data during the calibration process.
}
\seealso{
Other CMIP6: \code{\link{cmip6_Hector_inputs}},
  \code{\link{cmip6_center_scale}}, \code{\link{cmip6_ini}}
}
\concept{CMIP6}
\keyword{datasets}
