% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmip6_Hector_inputs}
\alias{cmip6_Hector_inputs}
\title{A list of dataframes containing CMIP6 emissions and concentrations for each scenario.}
\format{A list of data frames
\describe{
\item{driver.CMIP6scenario}{Is a dataframe containing a year, variable, value, and units column. All of the
contents of the data frame should be formatted as correct Hector inputs. }
}}
\usage{
cmip6_Hector_inputs
}
\description{
A list of datafame that contain the the CMIP6 emissions and concentration inputs for each
of the different CMIP6 scenarios labled by driver, either concentration or emissions.CMIP
scenario name.
}
\seealso{
Other CMIP6: \code{\link{cmip6_ini}}
}
\concept{CMIP6}
\keyword{datasets}
