% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cmip6_ini}
\alias{cmip6_ini}
\title{A mapping tibble of the CMIP6 scenario and relevant Hector CMIP5 ini file.}
\format{A data frame of the CMIP6 scenario name and the relvant hecotr CMIP5 ini file.
\describe{
\item{cmip6}{A column of the cmip6 scenario frame with the driver.CMIP6 nomenclature.}
\item{ini_name}{The name of a Hector CMIP5 ini file name, these ini files should live in the hector package
input direcotry}
}}
\usage{
cmip6_ini
}
\description{
A mapping tibble of the CMIP6 scenario name and the CMIP5 ini file. Because
Hector has not been updated with CMIP6 ini files this mapping tibble
can be used to generate Hector cores until the CMIP6 ini files have been
created.
}
\seealso{
Other CMIP6: \code{\link{cmip6_Hector_inputs}}
}
\concept{CMIP6}
\keyword{datasets}
