% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_calibration.R
\name{single_calibration_cmip6}
\alias{single_calibration_cmip6}
\title{Calibrate Hector to a single CMIP6 ESM.}
\usage{
single_calibration_cmip6(cmip6_runs, esm_data, normalize, initial_param,
  cmip_range = NULL, param_penalty = NULL, maxit = 500,
  n_parallel = NULL, showMessages = FALSE,
  intermediateOutput = FALSE)
}
\arguments{
\item{cmip6_runs}{A vector of the cmip6 scenario names to use to create Hector cores}

\item{esm_data}{A data frame of ESM data for a single model that contains the following columns, year, model, variable, experiment.}

\item{normalize}{A list of center and the scale values to use to noramlize the Hector and ESM output data.}

\item{initial_param}{A named vector of inital paramters to be optimized over.}

\item{cmip_range}{Default set to NULL and the make_minim_function will only look at the error between esm_data and Hector data.
But if cmip_range is set to a dataframe containing columns (variable, year, sig, lower, and upper) then the minimize function will also
minimize the value returned by the -log of the mesa function.}

\item{param_penalty}{Default is set to NULL but can be set to a function that will return a tibble of experiment / variable / value to penalize the
mean squared error for "unreal" paramter values.}

\item{maxit}{The max number of itterations for optim, default set to 500.}

\item{n_parallel}{The max number of cores to parallize the runs over,  unless sepcified will use the number of cores detected by \code{detectCores}.}

\item{showMessages}{Default set to FALSE, will supress Hector error messages.}

\item{intermediateOutput}{Default set to FALSE, but if set to TRUE will return the MSE for each variable / experiment / ensemble memeber instead over the over all MSE.}
}
\value{
An object returned by \code{optim}
}
\description{
Calibrate Hector to a single CMIP6 ESM.
}
